Update active_check.py
Normalized parameter names to filename and fout for the XML data file and the Excel output file, respectively. 
Added a few comments to some functions. 
Added the filename parameter to the functions without it so that changes to the target file name would properly filter through the function calls when using a name other than 'active_check.xml'. 
Modified pull_feed to return the filename string and formatted the parameter definitions for clarity. The returned filename string is not currently utilized, but since the function comments state that pull_feed returns the file name...